back_end/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pydantic>=2.0.0
langchain>=0.0.350
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
chromadb>=0.4.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pdf2image>=1.16.0
Pillow>=9.0.0
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.5.0
requests>=2.28.0
google-generativeai>=0.3.0
sentence-transformers